Fix Black / Flake8 issues

+++ mds/authent/public_api.py
@@ -6,24 +6,23 @@ from django.core.cache import cache
 from django.utils import timezone
 
 
-from mds.authent import models,generators
+from mds.authent import models, generators
+
 
 def get_long_lived_token(application, duration):
-    token_duration = datetime.timedelta(
-        seconds=duration
-    )
+    token_duration = datetime.timedelta(seconds=duration)
 
     token = generators.generate_jwt(application, token_duration, save=True)
     return token
 
-def revoke_long_lived_token(token):
-    stored_token = models.AccessToken.objects.filter(
-        token=token
-    ).last();
 
-    if (stored_token):
-        stored_token.revoked_after=timezone.now()
+def revoke_long_lived_token(token):
+    stored_token = models.AccessToken.objects.filter(token=token).last()
+
+    if stored_token:
+        stored_token.revoked_after = timezone.now()
         stored_token.save()
+
 
 def get_revocation_list() -> List[str]:
     """Get the list of revoked tokens.
@@ -53,6 +52,7 @@ def get_revocation_list() -> List[str]:
     cache.set(cache_key, token_ids, timeout=60)  # seconds
     return token_ids
 
+
 def create_application(name, owner=None, user=None, grant=None, scopes=None):
     grant = grant or models.Application.GRANT_CLIENT_CREDENTIALS
     return models.Application.objects.create(
@@ -67,11 +67,8 @@ def create_application(name, owner=None, user=None, grant=None, scopes=None):
 
 
 def revoke_application(owner_id):
-    models.Application.objects.filter(
-        owner=owner_id
-    ).update(scopes=[])
+    models.Application.objects.filter(owner=owner_id).update(scopes=[])
+
 
 def delete_application(owner_id):
-    models.Application.objects.filter(
-        owner=owner_id
-    ).delete()
+    models.Application.objects.filter(owner=owner_id).delete()
